// src/resources/hotel/searchHotelsUserInterfaces.ts
export enum searchPage {

    location = "//input[@id='MainContentPlaceHolder_SearchPanel_SearchPanelLayout_LocationComboBox_I']",
    select_location = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_LocationComboBox_DDD_L_LBI1T0",
    check_in = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_CheckInDateEdit_I",
    checkout = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_CheckOutDateEdit_I",
    rooms = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_RoomsComboBox_I",
    adults = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_AdultsSpinEdit_I",
    children = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_ChildrenSpinEdit_I",
    btn_search = "#MainContentPlaceHolder_SearchPanel_SearchPanelLayout_SearchButton_CD",
    btn_loginHeders = "#HeaderControl_Login_CD",
    email = "#HeaderControl_LogonControl_LoginFormLayout_txtEmail_I",
    password = "#HeaderControl_LogonControl_LoginFormLayout_txtPassword_I_CLND",
    code = "#HeaderControl_LogonControl_LoginFormLayout_Captcha_TB_I",
}